+++ bin/rails
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 
 ENGINE_ROOT = File.expand_path('../..', __FILE__)
-ENGINE_PATH = File.expand_path('../../lib/spree_gateway/engine', __FILE__)
+ENGINE_PATH = File.expand_path('../../lib/solidus_stripe/engine', __FILE__)
 
 require 'rails/all'
 require 'rails/engine/commands'
