+++ lib/solidus_stripe/version.rb
@@ -1,3 +1,3 @@
-module SolidusGateway
+module SolidusStripe
   VERSION = "1.2.0"
 end
